// src/generated/voiceList.ts
/* This file is auto-generated by scripts/generateSongsIndex.js */

export const VOICES = [
  {
    id: '1',
    name: 'alarm_clock_short_6402.mp3',
    uri: require('@assets/voices/alarm_clock_short_6402.mp3'),
    displayName: 'alarm-clock-short-6402',
    durationSec: 8,
    cover: null,
  },
  {
    id: '2',
    name: 'wake_up_33353.mp3',
    uri: require('@assets/voices/wake_up_33353.mp3'),
    displayName: 'wake-up-33353',
    durationSec: 20,
    cover: null,
  },
  {
    id: '3',
    name: 'wake_up_alarm_68067.mp3',
    uri: require('@assets/voices/wake_up_alarm_68067.mp3'),
    displayName: 'wake-up-alarm-68067',
    durationSec: 17,
    cover: null,
  },
];